requirements.txt
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
comtypes==1.1.14
cryptography==39.0.1
idna==3.4
PyAudio==0.2.13
pycparser==2.21
pycryptodome==3.17
pypiwin32==223
python-dotenv==0.21.1
pyttsx3==2.90
pywin32==305
requests==2.28.2
SpeechRecognition==3.9.0
urllib3==1.26.14
